+++ src/main/java/com/verdantartifice/thaumicwonders/common/registry/SoundsTW.java
@@ -7,6 +7,8 @@ import net.minecraft.util.SoundEvent;
 public class SoundsTW {
     public static final SoundEvent AVATAR_DEATH;
     public static final SoundEvent AVATAR_AMBIENT;
+    public static final SoundEvent INSPIRATION_ENGINE;
+    public static final SoundEvent MADNESS_ENGINE;
     public static final SoundEvent PRIMAL_DESTROYER_ANGERED;
     public static final SoundEvent VOID_BEACON_ACTIVATE;
     public static final SoundEvent VOID_BEACON_AMBIENT;
@@ -20,6 +22,8 @@ public class SoundsTW {
     static {
         AVATAR_AMBIENT = makeSoundEvent("avatar_ambient");
         AVATAR_DEATH = makeSoundEvent("avatar_death");
+        INSPIRATION_ENGINE = makeSoundEvent("inspiration_engine");
+        MADNESS_ENGINE = makeSoundEvent("madness_engine");
         PRIMAL_DESTROYER_ANGERED = makeSoundEvent("primal_destroyer_angered");
         VOID_BEACON_ACTIVATE = makeSoundEvent("vb_activate");
         VOID_BEACON_AMBIENT = makeSoundEvent("vb_ambient");