+++ src/main/java/com/verdantartifice/thaumicwonders/common/tiles/devices/TileMadnessEngine.java
@@ -1,5 +1,7 @@
 package com.verdantartifice.thaumicwonders.common.tiles.devices;
 
+import com.verdantartifice.thaumicwonders.common.registry.SoundsTW;
+import net.minecraft.util.SoundCategory;
 import thaumcraft.api.aspects.Aspect;
 
 public class TileMadnessEngine extends AbstractTileResearchEngine {
@@ -19,5 +21,12 @@ public class TileMadnessEngine extends AbstractTileResearchEngine {
     @Override
     public Aspect getAspect() {
         return Aspect.ELDRITCH;
+    }
+
+    @Override
+    public void playAmbientSound() {
+        if(this.tickCounter % 260 == 0) {
+            this.world.playSound(null, this.pos, SoundsTW.MADNESS_ENGINE, SoundCategory.BLOCKS, 0.7f, 1.0f);
+        }
     }
 }